data-analysis/marketplace-analytics/requirements.txt
dbt-core>=1.11.7,<2.0
dbt-duckdb>=1.10.1,<2.0
duckdb>=1.5.0,<2.0
kaggle>=2.0.0,<3.0
matplotlib>=3.10.8,<4.0
matplotlib-inline>=0.2.1,<1.0
numpy>=2.4.3,<3.0
pandas>=3.0.1,<4.0
pyarrow>=23.0.1,<24.0
scikit-learn>=1.8.0,,<2.0
scipy>=1.17.1,<2.0
python-dotenv>=1.0,<2.0